+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/TextContent.kt
@@ -5,8 +5,7 @@ import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.send.SendTextMessage
 import dev.inmo.tgbotapi.types.ChatIdentifier
 import dev.inmo.tgbotapi.types.MessageIdentifier
-import dev.inmo.tgbotapi.types.ParseMode.HTMLParseMode
-import dev.inmo.tgbotapi.types.ParseMode.ParseMode
+import dev.inmo.tgbotapi.types.ParseMode.*
 import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
 import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
 import dev.inmo.tgbotapi.types.message.content.abstracts.MessageContent
@@ -31,19 +30,24 @@ data class TextContent(
         replyMarkup
     )
 
+    @Deprecated(
+        "Useless due to fact that createResend currently use textSource and that will guarantee correct sending of message",
+        ReplaceWith("createResend")
+    )
     override fun createResends(
         chatId: ChatIdentifier,
         disableNotification: Boolean,
         replyToMessageId: MessageIdentifier?,
         allowSendingWithoutReply: Boolean?,
         replyMarkup: KeyboardMarkup?
-    ): List<Request<ContentMessage<TextContent>>> = createResends(
-        chatId,
-        disableNotification,
-        replyToMessageId,
-        allowSendingWithoutReply,
-        replyMarkup,
-        HTMLParseMode
+    ): List<Request<ContentMessage<TextContent>>> = listOf(
+        createResend(
+            chatId,
+            disableNotification,
+            replyToMessageId,
+            allowSendingWithoutReply,
+            replyMarkup
+        )
     )
 
     @Deprecated(
@@ -56,8 +60,6 @@ data class TextContent(
         replyToMessageId: MessageIdentifier?,
         allowSendingWithoutReply: Boolean?,
         replyMarkup: KeyboardMarkup?,
-        parseMode: ParseMode = HTMLParseMode
-    ): List<Request<ContentMessage<TextContent>>> = listOf(
-        createResend(chatId, disableNotification, replyToMessageId, allowSendingWithoutReply, replyMarkup)
-    )
+        parseMode: ParseMode = defaultParseMode
+    ): List<Request<ContentMessage<TextContent>>> = createResends(chatId, disableNotification, replyToMessageId, allowSendingWithoutReply, replyMarkup)
 }