+++ tgbotapi.core/src/commonTest/kotlin/dev/inmo/tgbotapi/types/MessageEntity/TextPartsCreatingTests.kt
@@ -1,8 +1,8 @@
 package dev.inmo.tgbotapi.types.MessageEntity
 
 import dev.inmo.tgbotapi.CommonAbstracts.justTextSources
-import dev.inmo.tgbotapi.utils.internal.toHtmlTexts
-import dev.inmo.tgbotapi.utils.internal.toMarkdownV2Texts
+import dev.inmo.tgbotapi.extensions.utils.formatting.toHtmlTexts
+import dev.inmo.tgbotapi.extensions.utils.formatting.toMarkdownV2Texts
 import kotlin.test.Test
 import kotlin.test.assertEquals