+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ParseMode/ParseMode.kt
@@ -35,12 +35,21 @@ typealias Markdown = MarkdownParseMode
 typealias MarkdownV2 = MarkdownV2ParseMode
 typealias HTML = HTMLParseMode
 
+/**
+ * This variable respects to default parse mode used in places like next:
+ *
+ * * [dev.inmo.tgbotapi.types.message.content.TextContent.createResends]
+ * *
+ */
+var defaultParseMode: ParseMode = HTML
+
 @Serializer(ParseMode::class)
 internal object ParseModeSerializerObject : KSerializer<ParseMode> {
     override fun deserialize(decoder: Decoder): ParseMode {
         return when (decoder.decodeString()) {
-            MarkdownParseMode.parseModeName -> MarkdownParseMode
-            HTMLParseMode.parseModeName -> HTMLParseMode
+            Markdown.parseModeName -> Markdown
+            MarkdownV2.parseModeName -> MarkdownV2
+            HTML.parseModeName -> HTML
             else -> throw IllegalArgumentException("Unknown parse mode")
         }
     }