+++ CHANGELOG.md
@@ -8,6 +8,8 @@
     * Several extensions for `TelegramBotAPI` like `retrieveAccumulatedUpdates` have been added as a solution for
     [#293](https://github.com/InsanusMokrassar/TelegramBotAPI/issues/293)
     * Links for `tg://user?id=<user_id>` have been updated ([#292](https://github.com/InsanusMokrassar/TelegramBotAPI/issues/292))
+    * All usages of captions or texts in resends and same things have been replaced with `textSources`
+    * Global `defaultParseMode` has been added ([#291](https://github.com/InsanusMokrassar/TelegramBotAPI/issues/291))
 
 ## 0.32.4